// frontend/src/app/api/models/about-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ThirdPartyComponentResponse } from '../models/third-party-component-response';
export interface AboutResponse {
  components: Array<ThirdPartyComponentResponse>;
  revision: (string | null);
  runtime: string;
  version: string;
}